requirements.txt
fastapi
uvicorn[standard]
itsdangerous
python-multipart
python-dotenv
Pillow
rawpy
numpy
